拉取回复、删除回复、删除评论回复的权限管理

+++ routes/api.php
@@ -89,12 +89,12 @@ Route::namespace('Api')->group(function (){
 
         //Eatest评论
         Route::post('eatest/{id}/comments','Eatest\CommentController@publish')->where(["id"=>"[0-9]+"])->middleware(['eatest.exist.check','comment.from.check']);
-
         Route::get('eatest/{id}/comments','Eatest\CommentController@get_list')->where(["id"=>"[0-9]+"])->middleware(['eatest.exist.check']);
-        Route::delete('eatest/{id}/comments','Eatest\CommentController@delete')->where(["id"=>"[0-9]+"])->middleware(['comment.exist.check']);
+        Route::delete('eatest/{id}/comments','Eatest\CommentController@delete')->where(["id"=>"[0-9]+"])->middleware(['comment.exist.check','comment.owner.check']);
         //Eatest评论回复
-        Route::post('eatest/{toId}/reply/{fromId}','Eatest\ReplyController@publish')->where(["toId"=>"[0-9]+","fromId"=>"[0-9]+"])->middleware('reply.exist.check');
-
+        Route::post('eatest/{toId}/reply/{fromId}','Eatest\ReplyController@publish')->where(["toId"=>"[0-9]+","fromId"=>"[0-9]+"])->middleware('reply.tofrom.check');
+        Route::get('eatest/{id}/reply','Eatest\ReplyController@get_list')->where(["id"=>"[0-9]+"])->middleware(['comment.exist.check']);
+        Route::delete('eatest/{id}/reply','Eatest\ReplyController@delete')->where(["id"=>"[0-9]+"])->middleware(['reply.exist.check','reply.owner.check']);
         /**User */
         Route::group(["middleware" => ['owner.check']], function () {
         //获取收藏列表
@@ -110,7 +110,7 @@ Route::namespace('Api')->group(function (){
         });
 
             Route::put('/notice/eatest/comments/{id}',"jwxt\NoticeController@eatest_comment_update")->where(["id" => "[0-9]+"])->middleware('comment.exist.check');
-            Route::put('/notice/eatest/reply/{id}',"jwxt\NoticeController@eatest_reply_update")->where(["id" => "[0-9]+"])->middleware('reply.id.check');
+            Route::put('/notice/eatest/reply/{id}',"jwxt\NoticeController@eatest_reply_update")->where(["id" => "[0-9]+"])->middleware('reply.exist.check');
 
         /**AssociationCode */
         //获取关联码
